also add a template for back-filling on old days.
this one, when you instantiate it, will use the current date.
The intent is you then go into the code and update the today object to
the date of where you've just inserted this.

uses the (month*30+day)/365 year percentage calculation since the day of
the year isn't readily available.

example: if you're looking at January 3, 2012, you'd change years to
2012, months to 1, and days to 3.

this will keep the life, year, and month bars frozen to january 3, 2012,
on this page.

if you happen to know a time of day (I have really old journal entries
where I just wrote one timestamped note on a day) you can hard-code
hour also (0-23) and it will freeze the day bar too.